requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp_socks==0.10.1
aiosignal==1.3.2
amqp==5.3.1
attrs==25.1.0
billiard==4.2.1
black==25.1.0
celery==5.4.0
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.1
frozenlist==1.5.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
matrix-nio==0.25.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
pycparser==2.22
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-socks==2.7.1
redis==5.2.1
referencing==0.36.2
rpds-py==0.22.3
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
unpaddedbase64==2.1.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
